requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
